worked on the wishlist controller and the User favourite controller and started working on the search Controller

+++ src/main/java/com/project/E_Commerce/dto/SearchRequest.java
@@ -15,6 +15,6 @@ public class SearchRequest {
     private Integer page;
     private List<String> sizes;
     private String sortBy;
-    private Double ratings;
+    private Integer ratings;
     private String color;
 }
